+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added PHPStorm meta for better suggestions
 - Added migration datetime to json output of migrate / rollback commands for future purpose
 - Added smallserial as primary in pqsql
+- Added method truncate() to migrations
 
 ### [1.2.0] - 2020-03-16
 #### Changed